Timetracking01/backend/requirements.txt
python-dotenv
flask 
flask-cors
pymysql
sqlalchemy  
flask_sqlalchemy  
Faker
psycopg2-binary
boto3
pytz